Small change to spack default package install path
In the current version, the default spack install results in
software being installed under /opt/ohpc/admin, which is
inaccessible to users. In order to preserve admin-only access
to the spack build tool, I'm suggesting the addition of a
second module for accessing spack-built applications.
This change sets spack to install software
under /opt/ohpc/pub/spack/, which will be
accessible cluster-wide. This also changes the default
path for spack modules, to
/opt/ohpc/pub/spack-%{version}-modules/
The spack-apps module is installed in
/opt/ohpc/pub/modulefiles/spack-apps/%{version} and points
to spack modulefiles placed in
/opt/ohpc/pub/spack-%{version}-modules/%{spack_arch_name}

The spack_arch_name part of the path is unfortunately built
fairly deeply into spack, and isn't possible to change via
the config files, but causes confusion in lmod with the
module names if it's not at the end of the MODULEPATH.

The distro version tests may need to be improved, if this
should support any other versions. I've also ignored aarch,
since spack isn't currently built in the arm-friendly OHPC
project.

Signed-off-by: Eric Coulter <jecoulte@iu.edu>